+++ src/main/java/com/wanted/legendkim/domain/board/boardcontroller/FreeBoardController.java
@@ -67,5 +67,5 @@ public class FreeBoardController {
 
         return "redirect:/freeboard/user/freeboard";
     }
-    //test
+
 }
