--- src/main/java/com/wanted/legendkim/domain/users/auth/config/SecurityConfig.java
@@ -1,61 +1,0 @@
-package com.wanted.legendkim.domain.users.auth.config;
-
-import com.wanted.legendkim.domain.users.auth.model.service.AuthService;
-import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-
-@Configuration
-@EnableWebSecurity
-public class SecurityConfig {
-
-    private final AuthService authService;
-
-    public SecurityConfig(AuthService authService) {
-        this.authService = authService;
-    }
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-    @Bean
-    public WebSecurityCustomizer webSecurityCustomizer() {
-        return web -> web.ignoring().requestMatchers(PathRequest.toStaticResources().atCommonLocations());
-    }
-
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        http
-                .userDetailsService(authService)
-                .authorizeHttpRequests(auth -> auth
-                        .requestMatchers("/", "/css/**", "/js/**", "/images/**", "/error").permitAll()
-                        .requestMatchers("/auth/**", "/user/signup").permitAll()
-                        .anyRequest().authenticated()
-                )
-                .exceptionHandling(conf -> conf
-                        .authenticationEntryPoint((request, response, authException) ->
-                                response.sendRedirect("/?error=login_required"))
-                )
-                .formLogin(form -> form
-                        .loginPage("/auth/login")
-                        .loginProcessingUrl("/login")
-                        .usernameParameter("email")
-                        .defaultSuccessUrl("/", true)
-                        .permitAll()
-                )
-                .logout(logout -> logout
-                        .logoutSuccessUrl("/")
-                        .permitAll()
-                );
-
-        return http.build();
-    }
-}
